Added Ravi's implementation of FSG support.
git-svn-id: https://cmusphinx.svn.sourceforge.net/svnroot/cmusphinx/trunk@3458 c60273ff-4a12-0410-8d58-82ceea6d4170

+++ sphinx2/src/libsphinx2/agc_emax.c
@@ -68,7 +68,7 @@ static float obs_max_sum = 0.0;
 
 int32 agcemax_set (double m)
 {
-    max = m;
+    max = (float) m;
     E_INFO("AGCEMax: %.2f\n", max);
     return 0;
 }
